Image Processing API and Image Processing page setup
Image processing page set up. Only a link for the page is set in the courses page - it will be changed with the further implementations.
Image processing API is implemented as a template but it will be tested and changed if necessary after the camera set up for the image processing page.